消除大部分的编译错误

+++ Alime_TransparentWidget.h
@@ -9,5 +9,5 @@ public:
     Alime_TransparentWidget(QWidget* parent = Q_NULLPTR);
     virtual bool nativeEvent(const QByteArray& eventType, void* message, long* result);
 private:
-    size_t boundaryWidth;
+    int boundaryWidth;
 };
